Convert a single departure to a tibble

This function is designed to parse the content returned by the interior steps of the departures functions.

departure_to_tibble(departure)

Arguments

departure

A departure, as a list, returned by the departures API call.

Value

A tibble consisting of the following columns:

stop_id

route_id

run_id (deprecated, use run_ref instead)

run_ref

direction_id

disruption_ids

scheduled_departure

estimated_departure

at_platform

platform_number

flags

departure_sequence