Convert a single route to a tibble

This function is designed to parse the content returned by the interior steps of the routes function. Service status information may be NA, depending on the API call that was used to populate the information.

route_to_tibble(route)

Arguments

route

A route, as a list, returned by the routes API call.

Value

A tibble of routes, with the following columns:

route_id

route_gtfs_id

route_name

route_type

route_type_description

route_number

geopath

service_status

service_status_timestamp